refactor(dashboard-tsr): split part-log-table row/header parts (#1392)
Move Badge, ThSort, Row + ExpandedRow and the sort constants into
part-log-table-parts.tsx; part-log-table now holds only the quick-filter
config and the stateful table shell. PartLogTable export unchanged.

Part of epic #1392
